# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Company API",
        "description": "API для управління компаніями з підтримкою версійності",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Локальний сервер розробки"
        }
    ],
    "paths": {
        "/api/company": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Створення/Оновлення компанії",
                "description": "Створює нову компанію або оновлює існуючу за ЄДРПОУ",
                "operationId": "storeCompany",
                "requestBody": {
                    "description": "Дані компанії",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "edrpou",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Назва компанії",
                                        "type": "string",
                                        "maxLength": 256,
                                        "example": "ТОВ Українська енергетична біржа"
                                    },
                                    "edrpou": {
                                        "description": "ЄДРПОУ компанії",
                                        "type": "string",
                                        "maxLength": 10,
                                        "example": "37027819"
                                    },
                                    "address": {
                                        "description": "Адреса компанії",
                                        "type": "string",
                                        "example": "01001, Україна, м. Київ, вул. Хрещатик, 44"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Компанію успішно створено - 'created'",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "created"
                                        },
                                        "company_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "version": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Компанію оновлено - 'updated' або виявлено дублікат - 'duplicate'",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "updated"
                                                },
                                                "company_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "version": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "duplicate"
                                                },
                                                "company_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "version": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Помилка валідації",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Назва компанії є обов'язковою (and 1 more error)"
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Назва компанії є обов'язковою"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Companies",
            "description": "Операції з компаніями"
        }
    ]
}